+++ src/Cdn/Cdn/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 
 ## Upcoming Release
+* Upgraded Microsoft.Cdn API version to 2026-04-01-preview.
+  - Added support for updating Azure Front Door rule sets.
+  - Added support for profile-level WAF policy and route associations on Azure Front Door security policies.
 * Fixed placeholder text in help examples for `New-AzFrontDoorCdnRuleEdgeActionObject`.
 
 ## Version 6.0.2